requirements.txt
aiogram>=3.13,<4.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-mistralai>=0.2.0
langchain-openai>=0.2.0
langgraph>=0.2.0
pandas>=2.2.0
openpyxl>=3.1.0
structlog>=24.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
chardet>=5.0.0
prometheus-client>=0.20.0
